made proto_locomotion used a config resource
